# Precompute indices