First commit:
Precompute looking up indices for all variables in `StackedTimeSolver Data`

This has an identical analysis as https://github.com/bankofcanada/ModelBaseEcon.jl/pull/41.
It is possible some aux function like this should be put into ModelBaseEcon

This takes solving the FRB-US model from 149ms to 83ms.

---------
Second commit:
disable an expensive assert

This has the same issue as the one worked around in https://github.com/bankofcanada/ModelBaseEcon.jl/pull/41.

It takes solving the FRB-US model from 83ms to 54ms

-----

The benchmarks have been made on top of https://github.com/bankofcanada/ModelBaseEcon.jl/pull/41