@bbejanov 2023-04-04
Thanks again @KristofferC!  This precomputation now is done in https://github.com/bankofcanada/ModelBaseEcon.jl/pull/35/commits/d9cdc09676270d8259ef99bcb3da58d9a1b565e2 
so here we can just use it.  Cheers